cleaned ui for create a trip
